fix: resolve critical issues and optimize rendering
- Add missing moves and collected fields to t_pacman structure
- Fix PACMAN/PLAYER constant inconsistency (now uses 'P' consistently)
- Fix size_t to int casting in ft_strlen comparisons
- Optimize rendering system to use images instead of pixel-by-pixel
- Add tile image loading and proper memory management
- Improve performance by reducing 1024 function calls per tile to 1
- Maintain full Norminette 4.1 compliance throughout

+++ src/map_loader.c
@@ -69,7 +69,7 @@ void calculate_map_dimensions(t_map *map)
     while (map->data[map->height])
         map->height++;
     
-    map->width = ft_strlen(map->data[0]);
+    map->width = (int)ft_strlen(map->data[0]);
 }
 
 void count_map_elements(t_map *map)
